Add race consistency function
Function checks for whether the race reporting of child is broadly consistent with the race of parents. This is complex given the large number of mulitracial categories. I generally err on the side of liberal inclusion here. The goal is simply to eliminate cases where there is a very clear disrepancy.