Add calculate_cond_means function
This function will create the coefficient table that I use to make the dotplot of effects. I am separating the creation of the table from the actual ggplot. Right now the function just uses some reasonable assumptions as baselines for the non-race variables. The function returns probabilities by default but there is an argument to return log-odds instead.